update: グラフ導入 月の日ごとで金額を登録
