+++ lib/view/calendar/calendar_page.dart
@@ -2,7 +2,7 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
 import 'package:foodcost/model/menu.dart';
 import 'package:foodcost/utils/authentication.dart';
-import 'package:foodcost/utils/firestore/foods.dart';
+import 'package:foodcost/utils/firestore/menus.dart';
 import 'package:foodcost/view/menu/create_menu_page.dart';
 import 'package:intl/intl.dart';
 import 'package:table_calendar/table_calendar.dart';
@@ -166,7 +166,7 @@ class _CalendarPageState extends State<CalendarPage> {
                       valueListenable: _selectedEvents,
                       builder: (context, value, _) {
                         return StreamBuilder<QuerySnapshot>(
-                            stream: FoodFirestore.menus
+                            stream: MenuFirestore.menus
                                 .where('user_id', isEqualTo: Authentication.myAccount!.id)
                                 .snapshots(),
                             builder: (context, snapshot) {
